requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130

# src/main.py
pyside6==6.10.1

# src/*.py
torch==2.9.1
torchvision==0.24.1
transformers==4.57.1

accelerate==1.12.0
easydict==1.13
filelock==3.20.0
pillow==12.0.0
numpy==2.3.5
matplotlib==3.10.7
einops==0.8.1
timm==1.0.22
addict==2.4.0
pymupdf==1.26.6

# ./get_model.py
huggingface_hub
hf_xet